poprawki do obliczania rozliczeń

+++ src/main/java/pl/janis/komornik/dto/DebtDto.java
@@ -1,4 +1,6 @@
 package pl.janis.komornik.dto;
 
-public record DebtDto(UserDto from, UserDto to, float amount) {
+import java.math.BigDecimal;
+
+public record DebtDto(UserDto from, UserDto to, BigDecimal amount) {
 }
